feat: add parse uuid pipe

+++ src/api/lesson/lesson.service.ts
@@ -7,6 +7,8 @@ import { Repository } from 'typeorm'
 import { ValidationException } from '@exceptions/validation.exception'
 import { ErrorCode } from '@constants/error-code.constant'
 import { ModuleEntity } from '@api/module/entities/module.entity'
+import { paginate } from '@utils/offset-pagination'
+import { ListLessonReqDto } from './dto/list-lesson.req.dto'
 
 @Injectable()
 export class LessonService {
@@ -37,12 +39,34 @@ export class LessonService {
     }
   }
 
-  findAll() {
-    return this.lessonRepository.find()
+  async findAll(reqDto: ListLessonReqDto) {
+    try {
+      const query = this.lessonRepository.createQueryBuilder('lessons').orderBy('lessons.createdAt', 'DESC')
+
+      const [lessons, metaDto] = await paginate<LessonEntity>(query, reqDto, {
+        skipCount: false,
+        takeAll: false
+      })
+
+      return {
+        lessons,
+        meta: metaDto
+      }
+    } catch (error) {
+      throw error
+    }
   }
 
-  findOne(id: string) {
-    return this.lessonRepository.findOne({ where: { lessonId: id } })
+  async findOne(id: string) {
+    try {
+      const lesson = await this.lessonRepository.findOne({ where: { lessonId: id } })
+      if (!lesson) {
+        throw new ValidationException(ErrorCode.L001, 'Lesson not found')
+      }
+      return lesson
+    } catch (error) {
+      throw error
+    }
   }
 
   async update(id: string, updateLessonDto: UpdateLessonDto) {
@@ -66,7 +90,15 @@ export class LessonService {
     }
   }
 
-  remove(id: string) {
-    return this.lessonRepository.delete(id)
+  async remove(id: string) {
+    try {
+      const lesson = await this.lessonRepository.findOne({ where: { lessonId: id } })
+      if (!lesson) {
+        throw new ValidationException(ErrorCode.L001, 'Lesson not found')
+      }
+      return this.lessonRepository.delete(id)
+    } catch (error) {
+      throw error
+    }
   }
 }
